[exercise-timer-big-ui] Split ExerciseTimer into it's own file
