+++ app/src/main/java/com/litus_animae/refitted/compose/exercise/Instruction.kt
@@ -10,7 +10,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.material.Button
 import androidx.compose.material.Card
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
@@ -19,15 +18,10 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.litus_animae.refitted.compose.state.ExerciseSetWithRecord
 import kotlinx.coroutines.Dispatchers
-import java.time.Instant
-import java.time.ZoneId
-import java.time.format.DateTimeFormatter
 
 
 @OptIn(ExperimentalFoundationApi::class)
@@ -67,7 +61,8 @@ private fun ColumnScope.ExerciseInstructions(
         Surface(
           Modifier
             .fillMaxWidth()
-            .padding(bottom = 5.dp)) {
+            .padding(bottom = 5.dp)
+        ) {
           Text(text = exerciseSet?.exerciseName ?: "", style = MaterialTheme.typography.h4)
         }
       }
@@ -83,36 +78,5 @@ private fun ColumnScope.ExerciseInstructions(
     }
   }
   // TODO this takes up space if not in use, looks bad with cards
-  val exerciseTimerRunning = rememberSaveable { mutableStateOf(false) }
-  val exerciseTimerMillis = rememberSaveable { mutableStateOf(0L) }
-  val isExerciseTimerRunning by exerciseTimerRunning
-  Timer(isExerciseTimerRunning,
-    millisToElapse = exerciseSet?.timeLimitMilliseconds ?: 0,
-    countDown = true,
-    animateTimer = false,
-    onUpdate = { exerciseTimerMillis.value = it }) { exerciseTimerRunning.value = false }
-  if (exerciseSet?.timeLimitMilliseconds != null) {
-    Row(Modifier.weight(1f)) {
-      Button(
-        onClick = {
-          if (!isExerciseTimerRunning) {
-            exerciseTimerMillis.value = exerciseSet.timeLimitMilliseconds
-          }
-          exerciseTimerRunning.value = !isExerciseTimerRunning
-        },
-        Modifier.fillMaxWidth()
-      ) {
-        val timerValue = Instant.ofEpochMilli(exerciseTimerMillis.value)
-          .atZone(ZoneId.systemDefault())
-          .toLocalTime()
-          .format(DateTimeFormatter.ofPattern("m:ss"))
-        val timerMax = Instant.ofEpochMilli(exerciseSet.timeLimitMilliseconds)
-          .atZone(ZoneId.systemDefault())
-          .toLocalTime()
-          .format(DateTimeFormatter.ofPattern("m:ss"))
-        if (isExerciseTimerRunning) Text("$timerValue remaining (click to stop)")
-        else Text("Start $timerMax exercise timer")
-      }
-    }
-  }
+  ExerciseTimer(timeLimitMilliseconds = exerciseSet?.timeLimitMilliseconds)
 }
